+++ src/components/ui/home/Preview/style.ts
